Moving tests.py to main directory
